refactor: test helpers Break up helper file, and add new `HyperviewMock` component to help with render tests
