--- test/render.test.ts
@@ -1,5 +1,0 @@
-import initStoryshots, { renderOnly } from '@storybook/addon-storyshots';
-
-initStoryshots({
-  test: renderOnly,
-});
